add inner join (name tbd) as docset transform
